requirements.txt
# Crowd Vision — Requirements
# Install with: pip install -r requirements.txt

ultralytics>=8.0.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
